+++ src/literalizer/languages/haskell.py
@@ -19,6 +19,7 @@ from literalizer._formatters.format_dates import (
     format_date_iso,
     format_datetime_epoch,
     format_datetime_iso,
+    format_time_iso,
 )
 from literalizer._formatters.format_entries import (
     format_bytes_base64,
@@ -1690,6 +1691,11 @@ class Haskell(metaclass=LanguageCls):
         ):
             return datetime_epoch_formatter(format_integer=self.format_integer)
         return self._date_fmts.format_datetime
+
+    @cached_property
+    def format_time(self) -> Callable[[datetime.time], str]:
+        """Callable that formats a time as a string literal."""
+        return format_time_iso
 
     @cached_property
     def format_float(self) -> Callable[[float], str]: