+++ src/literalizer/languages/sml.py
@@ -20,6 +20,7 @@ from literalizer._formatters.format_dates import (
     format_date_iso,
     format_datetime_epoch,
     format_datetime_iso,
+    format_time_iso,
 )
 from literalizer._formatters.format_entries import (
     declaration_formatter_ignoring_modifiers,
@@ -942,6 +943,11 @@ class Sml(metaclass=LanguageCls):
                 ),
             )
         return self.datetime_format
+
+    @cached_property
+    def format_time(self) -> Callable[[datetime.time], str]:
+        """Callable that formats a time as a string literal."""
+        return format_time_iso
 
     @cached_property
     def format_float(self) -> Callable[[float], str]: